+++ packages/tui/CHANGELOG.md
@@ -1,6 +1,6 @@
 # Changelog
 
-## [Unreleased]
+## [0.71.0] - 2026-04-30
 
 ### Fixed